+++ Package.swift
@@ -18,13 +18,12 @@ let package = Package(
     ],
     dependencies: [
         .package(url: "https://github.com/GetStream/stream-chat-swift.git", from: "4.22.0"),        
-        .package(url: "https://github.com/kean/Nuke.git", from: "10.0.0"),
-        .package(url: "https://github.com/kean/NukeUI.git", from: "0.7.0")
+        .package(url: "https://github.com/kean/Nuke.git", .exact("11.3.0"))
     ],
     targets: [
         .target(
             name: "StreamChatSwiftUI",
-            dependencies: [.product(name: "StreamChat", package: "stream-chat-swift"), "Nuke", "NukeUI"],
+            dependencies: [.product(name: "StreamChat", package: "stream-chat-swift"), "Nuke", .product(name: "NukeUI", package: "Nuke")],
             exclude: ["README.md", "Info.plist", "Generated/L10n_template.stencil"],
             resources: [.process("Resources")]
         )