+++ Sources/StreamChatSwiftUI/Utils/Common/NukeImageProcessor.swift
@@ -68,11 +68,11 @@ extension ImageProcessors {
             return size
         }
 
-        private let sizeProvider: () -> CGSize
+        private let sizeProvider: @Sendable() -> CGSize
         
         /// Initializes the processor with size providing closure.
         /// - Parameter sizeProvider: Closure to obtain size after the image is loaded.
-        public init(sizeProvider: @escaping () -> CGSize) {
+        public init(sizeProvider: @escaping @Sendable() -> CGSize) {
             self.sizeProvider = sizeProvider
         }