+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 # Upcoming
 
 ### 🔄 Changed
+- Updated Nuke dependency to 11.3.0 for SPM
+- Removed NukeUI dependency for SPM (now part of Nuke)
 
 # [4.22.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.22.0)
 _September 27, 2022_