+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### 🔄 Changed
 - Public init for `DefaultChannelListHeaderModifier`
+- Updated Nuke dependency
 
 ### 🐞 Fixed
 - Scroll to bottom when return key is pressed in the composer input view