+++ CHANGELOG.md
@@ -3,6 +3,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🔄 Changed
+
+# [4.24.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/4.24.0)
+_November 16, 2022_
+
 ### ✅ Added
 - Support for specifying whether ChatChannelListView is embedded in a NavigationView with `embedInNavigationView`